Finished Uploads and auth integration
Added the rest of the decryption/encryption functions, updated existing ones to fix some issues. Added google OAuth with AppWrite. Updated Appwrite implementation. Added API routes for uploading files, and one for fetching the encrypted file. created pages and components for the pages of viewing images. And routes for signup